sem-3/tubes-algeo-2/src/backend/requirements.txt
fastapi==0.122.1
filetype==1.2.0
nltk==3.9.2
numpy==2.3.5
Pillow==12.0.0
python-multipart==0.0.20
scipy==1.16.3
uvicorn==0.38.0
